requirements.txt
langchain
langchain_community
langchain_groq
faiss-cpu
pypdf
tiktoken
python-dotenv
streamlit
sentence-transformer
pypdf2